cctm-0.0.0.1.0: First three include class renames.
